clippy fixes

+++ src/time/cds.rs
@@ -338,9 +338,8 @@ impl<ProvidesDaysLen: ProvidesDaysLength> TimeProvider<ProvidesDaysLen> {
             unix_seconds: 0,
             submillis_precision: None,
         };
-        let unix_days_seconds =
-            ccsds_to_unix_days(ccsds_days.into()) as i64 * SECONDS_PER_DAY as i64;
-        provider.setup(unix_days_seconds as i64, ms_of_day.into());
+        let unix_days_seconds = ccsds_to_unix_days(ccsds_days.into()) * SECONDS_PER_DAY as i64;
+        provider.setup(unix_days_seconds, ms_of_day.into());
         Ok(provider)
     }
 
@@ -552,7 +551,7 @@ impl<ProvidesDaysLen: ProvidesDaysLength> CcsdsTimeProvider for TimeProvider<Pro
     }
 
     fn date_time(&self) -> Option<DateTime<Utc>> {
-        self.calc_date_time((self.ms_of_day % 1000) as u32)
+        self.calc_date_time(self.ms_of_day % 1000)
     }
 }
